refactor: 调整 html 标签，以更好的支持 a11y
