+++ src/templates/Layout.js
@@ -1,12 +1,12 @@
-import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
 import { Link } from 'gatsby';
-import Toggle from './ThemeToggle';
-
 import 'prismjs/themes/prism-okaidia.css';
+import React, { useEffect, useState } from 'react';
+import styled from 'styled-components';
 import { rhythm, scale } from '../utils/typography';
+import Bio from './Bio';
+import Toggle from './components/ThemeToggle';
 
-const PostContainer = styled.main`
+const PostContainer = styled.section`
   color: var(--textNormal);
   background: var(--bg);
   transition: color 0.3s ease-out, background 0.3s ease-out;
@@ -85,7 +85,7 @@ const Layout = ({ location, title, children }) => {
   }
   return (
     <PostContainer>
-      <article
+      <section
         style={{
           marginLeft: 'auto',
           marginRight: 'auto',
@@ -109,8 +109,9 @@ const Layout = ({ location, title, children }) => {
             }}
           />
         </Header>
+        <Bio />
         {children}
-      </article>
+      </section>
     </PostContainer>
   );
 };
