+++ src/templates/avatar.jpg
